VS Code: clean up
