+++ vscode/src/tree-sitter/query-sdk.test.ts
@@ -1,10 +1,9 @@
 import { afterEach, describe, expect, it } from 'vitest'
 
-import { initTreeSitterParser } from '../completions/test-helpers'
-
 import { SupportedLanguage } from './grammars'
 import { resetParsersCache } from './parser'
 import { getDocumentQuerySDK } from './query-sdk'
+import { initTreeSitterParser } from './test-helpers'
 
 describe('getDocumentQuerySDK', () => {
     afterEach(() => {
