+++ vscode/src/tree-sitter/query-tests/node-at-cursor-and-parents.test.ts
@@ -1,9 +1,9 @@
 import { beforeAll, describe, it } from 'vitest'
 
-import { initTreeSitterParser } from '../../completions/test-helpers'
 import { getNodeAtCursorAndParents } from '../ast-getters'
 import { SupportedLanguage } from '../grammars'
 import { getDocumentQuerySDK } from '../query-sdk'
+import { initTreeSitterParser } from '../test-helpers'
 
 import { annotateAndMatchSnapshot } from './annotate-and-match-snapshot'
